modificando compras

+++ database/migrations/2020_02_06_225459_create_compras_table.php
@@ -16,7 +16,7 @@ class CreateComprasTable extends Migration
         Schema::create('compras', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('ptoventa');
-            $table->integer('numcompra');
+            $table->bigInteger('numcompra');
             $table->string('fecha');
             $table->string('fechaCompra')->nullable();
             $table->decimal('recargo', 12, 2);
